Users now logged in automatically successfully signing up
